1.0.0/requirements.txt
Flask>=0.9
azure-iot-device~=2.1.0